minor formatting changes
